requirements.txt
streamlit
keras==3.13.2
tensorflow-cpu==2.19.0
numpy<2.0.0
opencv-python-headless
Pillow
minio
psycopg2-binary
python-dotenv
plotly
supabase